requirements.txt
langchain_astradb
pandas
langchain_google_genai
fastapi
uvicorn
python-multipart
jinja2
python-dotenv
langchain
langchain_core

-e .